requirements.txt
streamlit==1.28.0
langchain==0.0.319
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
PyMuPDF==1.23.5
pandas==2.1.1
plotly==5.18.0
google-generativeai==0.3.1
python-dotenv==1.0.0
matplotlib==3.8.0
seaborn==0.13.0
numpy==1.26.1